Решение лабораторной Task01
