--- check.txt
@@ -1,10 +1,0 @@
-# Позитивные тесты:
-(3, 3, 3)  -> equilateral
-(3, 4, 3)  -> isosceles
-(3, 4, 5)  -> nonequilateral
-
-# Негативные тесты:
-(1, 1, 2)  -> Исключение IncorrectTriangleSides
-(0, 1, 1)  -> Исключение IncorrectTriangleSides
-(-3, 3, 3) -> Исключение IncorrectTriangleSides
-(5, 10, 1) -> Исключение IncorrectTriangleSides
