--- equation.txt
@@ -1,12 +1,0 @@
-from quad_solver import solve_quad
-
-1)  print(solve_quadratic(1, 0, -4))  # Ожидаемый вывод: [-2.0, 2.0]
-2)  print(solve_quadratic(1, -2, 1))  # Ожидаемый вывод: [1.0]
-3)  print(solve_quadratic(1, 0, 4))   # Ожидаемый вывод: []
-4)  print(solve_quadratic(0, 2, -4))  # Ожидаемый вывод: [2.0]
-5)  print(solve_quadratic(0, 0, 3))   # Ожидаемый вывод: []
-6)  print(solve_quadratic(0, 0, 0))   # Ожидаемый вывод: ["infinite solutions"]
-7)  print(solve_quadratic(1, -3, 2))  # Ожидаемый вывод: [1.0, 2.0]
-8)  print(solve_quadratic(1, 2, 1))   # Ожидаемый вывод: [-1.0]
-9)  print(solve_quadratic(2, 5, 2))   # Ожидаемый вывод: [-2.0, -0.5]
-10) print(solve_quadratic(4, -12, 9)) # Ожидаемый вывод: [1.5]
